chore: move enum to types for consistency
