+++ src/shillelagh/adapters/api/gsheets/lib.py
@@ -29,11 +29,11 @@ from shillelagh.adapters.api.gsheets.typing import UrlArgs
 from shillelagh.exceptions import ProgrammingError
 from shillelagh.fields import Field
 from shillelagh.fields import Float
-from shillelagh.fields import Order
 from shillelagh.fields import String
 from shillelagh.filters import Equal
 from shillelagh.filters import Filter
 from shillelagh.filters import Range
+from shillelagh.types import Order
 from shillelagh.typing import Row
