+++ src/shillelagh/adapters/api/weatherapi.py
@@ -17,10 +17,10 @@ from shillelagh.fields import Boolean
 from shillelagh.fields import DateTime
 from shillelagh.fields import Float
 from shillelagh.fields import Integer
-from shillelagh.fields import Order
 from shillelagh.fields import String
 from shillelagh.filters import Filter
 from shillelagh.filters import Range
+from shillelagh.types import Order
 from shillelagh.typing import RequestedOrder
 from shillelagh.typing import Row
