+++ src/shillelagh/fields.py
@@ -1,6 +1,5 @@
 import datetime
 from distutils.util import strtobool
-from enum import Enum
 from typing import Any
 from typing import Callable
 from typing import cast
@@ -10,21 +9,7 @@ from typing import Type
 
 import dateutil.parser
 from shillelagh.filters import Filter
-
-
-class Order(Enum):
-    # Use ASCENDING/DESCENDING when you have static data with 1+
-    # columns pre-sorted. All other columns should have Order.NONE.
-    ASCENDING = "ascending"
-    DESCENDING = "descending"
-
-    # Use NONE when you can't or don't want to sort the data. Sqlite
-    # will then sort the provided data according to the query.
-    NONE = "none"
-
-    # Use ANY when the column can be sorted in any order. Usually
-    # all other columns will also have Order.ANY.
-    ANY = "any"
+from shillelagh.types import Order
 
 
 class Field:
