+++ src/shillelagh/types.py
@@ -1,10 +1,25 @@
 import datetime
 import inspect
 import time
+from enum import Enum
 from typing import Any
 
-from shillelagh.fields import Field
 from typing_extensions import Literal
+
+
+class Order(Enum):
+    # Use ASCENDING/DESCENDING when you have static data with 1+
+    # columns pre-sorted. All other columns should have Order.NONE.
+    ASCENDING = "ascending"
+    DESCENDING = "descending"
+
+    # Use NONE when you can't or don't want to sort the data. Sqlite
+    # will then sort the provided data according to the query.
+    NONE = "none"
+
+    # Use ANY when the column can be sorted in any order. Usually
+    # all other columns will also have Order.ANY.
+    ANY = "any"
 
 
 class DBAPIType:
@@ -12,7 +27,7 @@ class DBAPIType:
         self.name = name
 
     def __eq__(self, other: Any) -> bool:
-        if inspect.isclass(other) and issubclass(other, Field):
+        if inspect.isclass(other) and hasattr(other, "db_api_type"):
             return bool(self.name == other.db_api_type)
 
         return NotImplemented
