+++ tests/adapters/api/gsheets/test_adapter.py
@@ -14,9 +14,9 @@ from shillelagh.backends.apsw.db import connect
 from shillelagh.exceptions import InternalError
 from shillelagh.exceptions import ProgrammingError
 from shillelagh.fields import Float
-from shillelagh.fields import Order
 from shillelagh.fields import String
 from shillelagh.filters import Equal
+from shillelagh.types import Order
 
 from ....fakes import FakeAdapter
 from ....fakes import FakeEntryPoint
