+++ tests/backends/apsw/test_vt.py
@@ -15,12 +15,12 @@ from shillelagh.exceptions import ProgrammingError
 from shillelagh.fields import Field
 from shillelagh.fields import Float
 from shillelagh.fields import Integer
-from shillelagh.fields import Order
 from shillelagh.fields import String
 from shillelagh.fields import type_map
 from shillelagh.filters import Equal
 from shillelagh.filters import Filter
 from shillelagh.filters import Range
+from shillelagh.types import Order
 from shillelagh.typing import Row
 
 from ...fakes import FakeAdapter
