+++ tests/test_lib.py
@@ -3,7 +3,6 @@ from shillelagh.exceptions import ImpossibleFilterError
 from shillelagh.exceptions import ProgrammingError
 from shillelagh.fields import Float
 from shillelagh.fields import Integer
-from shillelagh.fields import Order
 from shillelagh.fields import String
 from shillelagh.filters import Equal
 from shillelagh.filters import Impossible
@@ -19,6 +18,7 @@ from shillelagh.lib import RowIDManager
 from shillelagh.lib import serialize
 from shillelagh.lib import unquote
 from shillelagh.lib import update_order
+from shillelagh.types import Order
 
 
 def test_row_id_manager_empty_range():
